주소 input required (#171)
